Rename existing wrapper, to match contract name
